+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - layers:
     - Errorhandling for secure services.
 ### Changed
+- Layers:
+  - Sta: setting the default style of historical features if there are no style defined.
 
 ### Deprecated